# Remote Write V2 Gogoproto Replacement with Vtprotobuf