--- a/.gitignore
+++ b/.gitignore
@@ -31,3 +31,7 @@ npm_licenses.tar.bz2
 
 # Ignore parser debug
 y.output
+
+# editors
+**/.idea/*
+**/.vscode/*