@francoposa 2025-02-12
## Still TODO at this point:

* Adjust protobuf build for `buf.build`
* Investigate CPU Time diff on `BenchmarkBuildV2WriteRequest`
* Some more extensive benchmarking?