@francoposa 2025-02-12
## Update: Pprof for CPU  Time on `BenchmarkBuildV2WriteRequest`

The profiles on each branch are zipped up here:
[CPU-profile-BenchmarkBuildV2WriteRequest.zip](https://github.com/user-attachments/files/18773209/CPU-profile-BenchmarkBuildV2WriteRequest.zip)

The difference seems to be in encoding LabelsRefs, as using the vtprotobuf code without any manual patching lost this optimization from `prompb/io/prometheus/write/v2/custom.go`:
```golang
if len(m.LabelsRefs) > 0 {
		// This is the trick: encode the varints in reverse order to make it easier
		// to do it in place. Then reverse the whole thing.
```

It would be a question for the maintainers whether continuing to maintain this patch is worth it.