@francoposa 2025-02-13
thanks @bwplotka! Will take another round at this when I am back from vacation

Re: Pooling - Pooling timeseries and samples performed better on these benchmarks than vtprotobuf without any pooling. I will evaluate those again individually, as well as see if prombench shows any different characteristics.

Re: Measuring Allocs - BuildWriteRequestV2 shows slower CPU time even with 0 allocs in most of the cases. Will chase down the time spent in Size more here, plus the effects of the LabelsRef patch optimization.