@francoposa 2025-04-25
> Hello from the bug-scrub! @francoposa do you think you will come back to this?

@bboreham I think based on the [conversation in CNCF Slack](https://cloud-native.slack.com/archives/C03J794L0BV/p1740132588920689?thread_ts=1740077322.830869&cid=C03J794L0BV) and my own research on the Mimir side that there is little confidence that we can ever really get the performance we want without the non-nullables either from Gogoproto or from figuring something else out / creating a separate generator for Prometheus' use cases.